Add random seeds to the training script
This adds random seeds for NumPy and TensorFlow to the training script
to allow for reproducible experiments.  This work uncovered a potential
issue within TensorFlow in which non-determinism can occur with the
`tf.data.Dataset.map` function when `num_parallel_calls` > 1 and the
function body contains ops that make use of randomness, despite graph
and op-level seeds.  An issue has been opened, and hopefully
improvements will be made.  In the meantime, determinism can be achieved
by setting the `threads` to 1, at the cost of runtime performance.