Fix normalization function NumPy datatype issue
Currently, we have a generic normalization function that can be used
with either TensorFlow or NumPy due to duck typing.  During training and
evaluation, we use TensorFlow ops for the full pipeline, including the
normalization step of the preprocessing.  This pipeline reads in images
into a float32 format prior to running normalization and the rest of the
pipeline.  Therefore, the normalization ops will correctly yield float32
outputs.  However, when the normalization function is called with NumPy
inputs, such as during our prediction script, the normalization function
incorrectly yields float64 ouputs for ImageNet-style normalization.
This is due to the ImageNet means literal array being implicitly
converted to a float64 array instead of a float32 array.  Once that
happens, the mean subtraction op will yield a float64 output, regardless
of the input datatype.  Testing reveals that this causes significant
differences in the output of the normalization function, as well as the
downstream predictions of a full model.

This updates the normalization function to explicitly create the
ImageNet means as a float32 tensor.  Testing confirms that no change is
experienced for TensorFlow inputs, while NumPy inputs now yield results
equivalent to that with TensorFlow inputs.